Move Unit Organization into Unit Info and remove unused tab panel
