+++ src/app/layout/App.js
@@ -8,7 +8,6 @@ import TabBarContainer from "features/tabs/TabBarContainer";
 import UnitInfo from "features/unitInfo/UnitInfo";
 import Pilots from "features/pilots/Pilots";
 import Mechs from "features/mechs/Mechs";
-import UnitOrganization from "features/unitOrganization/UnitOrganization";
 import Tools from "features/tools/Tools";
 import ModalManager from "features/modals/ModalManager";
 import ContextMenuManager from "features/contextMenus/ContextMenuManager";
@@ -21,7 +20,6 @@ class App extends Component {
             {name : "unitInfo", label : "Unit Info", component : UnitInfo,},
             {name : "pilots", label : "Pilots", component : Pilots,},
             {name : "mechs", label : "Mechs", component : Mechs,},
-            {name : "unitOrganization", label : "Unit Organization", component : UnitOrganization},
             {name : "tools", label : "Tools", component : Tools},
         ];
