+++ src/app/schema/schema.js
@@ -5,7 +5,7 @@ import MechDesign from "features/mechs/MechDesign";
 import Mech from "features/mechs/Mech";
 import Unit from "features/unitInfo/Unit";
 import Faction from "features/unitInfo/Faction";
-import Lance from "features/unitOrganization/Lance";
+import Lance from "features/unitInfo/Lance";
 
 const orm = new ORM();
 orm.register(Pilot, MechDesign, Mech, Unit, Faction, Lance);
