--- src/features/unitOrganization/UnitOrganization/UnitOrganization.jsx
@@ -1,28 +1,0 @@
-import React from "react";
-
-import {
-    Grid,
-    Segment,
-    Header,
-} from "semantic-ui-react";
-
-import UnitOrganizationTree from "../UnitOrganizationTree";
-
-
-const UnitOrganization = () => {
-    return (
-        <Segment>
-            <Grid>
-                <Grid.Column width={10}>
-                    <Header as="h3">Unit Table of Organization</Header>
-                    <Segment>
-                        <UnitOrganizationTree />
-                    </Segment>
-
-                </Grid.Column>
-            </Grid>
-        </Segment>
-    )
-}
-
-export default UnitOrganization
