--- src/features/unitOrganization/UnitOrganization/index.js
@@ -1,1 +1,0 @@
-export {default} from "./UnitOrganization";
